requirements.txt
python-dotenv == 1.0.0
pylint == 2.17.4
requests == 2.31.0 